fix clang-tidy

+++ simulation_adapter/src/simulation_adapter.cpp
@@ -345,7 +345,7 @@ void SimulationAdapter::egoDataCallback(const pm::EgoData::ConstSharedPtr& msg) 
     for (int i = 0; i < n; i++) {
       // update header stamp
       object_state.header = trajectory_planned_snapshot.header;
-      const double time = static_cast<double>(trajectory_planning_msgs::trajectory_access::getT(trajectory_planned_snapshot, i));
+      const double time = trajectory_planning_msgs::trajectory_access::getT(trajectory_planned_snapshot, i);
       object_state.header.stamp = rclcpp::Time(trajectory_planned_snapshot.header.stamp) + rclcpp::Duration::from_seconds(time);
 
       perception_msgs::object_access::setX(object_state,
